Track multiple rcpt in the message
This also tracks the from as if there could be multiple from addresses,
but that seems fine.  I may change how this is layed out, but at least
we're no longer losing information.

This also stops putting message data into the "state" except the
message-id so we can look up the current message.